Refactor soWatcher client
