--- pkg/network/usm/sharedlibraries/testutil/sowatcher_client/.gitignore
@@ -1,1 +1,0 @@
-sowatcher_client
